requirements.txt
# FastAPI dependencies
fastapi==0.109.0
uvicorn==0.23.2
pydantic==2.5.2
pydantic-settings==2.1.0
python-multipart==0.0.19
python-jose[cryptography]==3.4.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
httpx==0.25.0

# AWS dependencies
boto3==1.34.43
botocore==1.34.43
aws-secretsmanager-caching==1.1.1.5

# Database
sqlalchemy==2.0.22
alembic==1.12.0
psycopg2-binary==2.9.10

# Testing
pytest==7.4.2
pytest-asyncio==0.21.1

# Development
black==24.3.0
isort==5.12.0
flake8==6.1.0
mypy==1.6.1